project_2mm/project_2mm/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.1
defusedxml==0.7.1
distlib==0.3.6
dj-rest-auth==4.0.1
Django==4.2.1
django-allauth==0.54.0
django-cors-headers==4.2.0
django-embed-video==1.4.8
django-phonenumber-field==7.1.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
drf-yasg==1.21.7
filelock==3.10.4
idna==3.4
inflection==0.5.1
install==1.3.5
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.1
phonenumbers==8.13.18
Pillow==9.5.0
pipenv==2023.3.20
platformdirs==3.1.1
pycparser==2.21
PyJWT==2.7.0
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
simplejson==3.19.1
six==1.16.0
social-auth-core==4.4.2
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
virtualenv==20.21.0
virtualenv-clone==0.5.7
